refactor(py3createtorrent.py): update Config class initialization, and use native type hinting
This commit refactors the `Config` class initialization in the
`py3createtorrent.py` file. The `path` parameter is now optional and can
be passed during initialization. The `advertise` attribute is now of
type `Optional[bool]`. The `best_trackers_url` attribute is of type
`str`.

Additionally, the commit updates type annotations in the `split_path`
function and the `main` function. The `input_path` and `trackers`
variables now have explicit type annotations. The `metainfo` variable in
the `main` function now has a type hint of `Dict[str, Any]`.